Fixed building of libinitng.a
The target directory for the object files wasn't set correctly, and thus, if
the file "foo.c" occurred multiple times with different content in several
subdirectories, only one "foo.o" was created in the core/ directory, with
subsequent object files overwriting the previous ones. libinitng.a contained
an incomplete set of functions, and linking against it was impossible.

Just gotta fix the warnings in the other files now. :^)